backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
alembic>=1.13
pydantic>=2.6
pydantic-settings>=2.2
email-validator>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart>=0.0.9
Jinja2>=3.1
PyMySQL>=1.1
